Update patch set 5
Patch Set 5:

(3 comments)

We should test this with some app that has cleartextallowed explicitly set to false to make sure the AOSP functionality isn't broken given that we've modified this code.

A good example is the Updater, it has cleartextTrafficPermitted=false set, so if you change the string to a http url for testing it shouldn't work and complain in logcat.

Patch-set: 5
Reviewer: Gerrit User 1000000 <1000000@35266860-6252-469c-a67d-185064479fc0>