Return ECONNREFUSED error code when a process is not allowed networking
Libcore throws a SecurityException to remind app developers to declare the INTERNET permission in the app's manifest if it attempts to do networking without it being present (see https://cs.android.com/android/platform/superproject/+/master:libcore/ojluni/src/main/java/java/net/Inet6AddressImpl.java;l=150;drc=c46e557aedf10d58aeb8c356696981ad03c666b1)
When an app's networking is restricted, the INET GID is not reported to netd in order to block the creation of AF_INET sockets. Returning the EPERM error code causes this exception to be thrown even if an app does declare the permission and it is granted. This patch replaces the error code with ECONNREFUSED.

Change-Id: Ie78e3a064ff09ac67863601034e8ee9d66785b06