Add drop rule and routes for virtual networks
When a VPN file descriptor is closed, its routes are instantly removed,
by design of the TUN/TAP kernel module. This makes it impossible for
the system to react in real time, before outbound traffic falls through
to other interfaces. To make this possible, for all route modifications
involving a virtual network, also modify the routes in a separate table
designated to drop the traffic by using the dummy interface instead.
Consult this new table via a new rule placed below all VPN UID range
rules. When a VPN disconnects, there is now an opportunity to react in
the onInterfaceRemoved event of netd listeners, before any traffic
previously destined for the VPN can proceed.

Change-Id: I2aa94ae6bcb5c28eb5f7cc4f6ff1b9a3cb95468b