Do not add VPN local exclusion rules
Prevent UIDs on a VPN from accessing local network resources when they
are otherwise not allowed to access those resources. Without this, when connected to a VPN, apps that are disallowed from accessing a Wi-Fi network will become able to access the Wi-Fi network's local resources, despite not being able to access those resources when not connected to
a VPN. This change removes the rule that makes this bypass possible.

Change-Id: I9975b5ab1306ee86863979d1fe73203799cce648