Stop setting netId for bypassable VPNs
Code comments claim that a socket's netId needs to be specified for
bypassable VPNs in order for them to have any traffic at all, but this
does not appear to reflect reality today, as a simple test will show
that such VPNs are still usable even without this.

This change resolves an issue resulting from recent firewall changes
that prevents UIDs of bypassable VPNs from accessing other networks,
even when they are allowed to do so.

Issue: calyxos#1650
Change-Id: I18edc8659750044534c9bea5ed49eddbcea89378