Prevent DNS traffic from bypassing lockdown VPNs
Only allow bypassing a lockdown VPN during DNS lookups if the UID can
protect sockets from the VPN, which (as always) requires special
permission. This prevents DNS leaks when the VPN is being reconfigured.

Additionally, as a consequence of this change, VPNs which do not specify
DNS servers of their own may be unable to resolve names in lockdown mode
if the system DNS is on a LAN and thus not accessible over the VPN.
This prevents unexpected and on-going cleartext lookups in lockdown.

Bug: 337961996
Issue: calyxos#2555
Change-Id: Id1d27a6e28f4cd126ef7e3099412bf859c7a6ad1