Add Fwmark flag to force use of the given netId
Provide a way to ensure that traffic, such as DNS validation traffic,
will only traverse the netId specified and not fall through to any
other networks. If the network disappears, the destination should be
deemed unreachable. The network must *also* be explicitly selected,
or the destination will be unreachable.

Issue: calyxos#2680
Change-Id: I0e5cb87852086816edf22cdbb2e129fa3f40b552