Add Fwmark flag to prevent fallthrough
Provide an Fwmark flag, enforceNetId, to ensure that traffic - such as
DNS validation traffic - does not fall through to other routing rules
if not matched by explicit network rules; instead, the processing stops
with a "Network is unreachable" error. The Fwmark must also specify the
explicitlySelected flag or it will likely also be affected.

Issue: calyxos#2680
Change-Id: I0e5cb87852086816edf22cdbb2e129fa3f40b552